requirements.txt
amqp==2.1.4
anyjson==0.3.3
appdirs==1.4.0
billiard==3.3.0.23
celery==3.1.25
CouchDB==1.1
cycler==0.10.0
Django==1.10.5
django-celery==3.2.1
djangorestframework==3.5.4
imutils==0.3.6
kombu==3.0.37
mahotas==1.4.1
matplotlib==1.5.1
numpy==1.11.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.1.10
python-dateutil==2.5.3
pytz==2016.10
requests==2.13.0
scikit-learn==0.17.1
scipy==0.18.0
six==1.10.0
vine==1.1.3
